// openapi/dres/model/queryEvent.ts
/**
 * DRES Client API
 * Client API for DRES (Distributed Retrieval Evaluation Server), Version 2.0.3
 *
 * The version of the OpenAPI document: 2.0.3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { QueryEventCategory } from './queryEventCategory';


export interface QueryEvent { 
    timestamp: number;
    category: QueryEventCategory;
    type: string;
    value: string;
}